refactor: migrate BGP CRDs from Cosmos to datum-cloud/network
Cosmos is being phased out; its BGP CRD types (BGPRouter, BGPPeer,
BGPAdvertisement, BGPPolicy, BGPVRFInstance) are field-for-field
identical in the new datum-cloud/network module, so this is a clean
cutover of the import path, API group, and CRD source rather than a
logic change. VPC/VPCAttachment CRDs stay on Cosmos since they belong
to the separate companion VPC operator.

- Swap the bgpv1alpha1 import from go.miloapis.com/cosmos to
  go.datum.net/network across all controllers, reconcile, model, cni,
  and gc packages
- Move RBAC and the galactic-router RBAC pre-flight check from the
  bgp.miloapis.com API group to network.datumapis.com
- Point deploy-system.sh at datum-cloud/network for BGP CRD manifests,
  keeping VPC CRDs on Cosmos
- Update containerlab sample CRs and docs (ARCHITECTURE.md, AGENTS.md)
  to reflect the new dependency

closes #199
