+++ ARCHITECTURE.md
@@ -2,7 +2,7 @@
 
 > Galactic is the SRv6 data plane for multi-cloud VPC networking, deployed as two
 > binaries on each Kubernetes node: a CNI plugin that attaches containers to VPC
-> networks, and a router that reconciles Cosmos BGP CRDs and drives an embedded
+> networks, and a router that reconciles BGP CRDs and drives an embedded
 > GoBGP server to distribute EVPN (L2VPN/EVPN AFI/SAFI) paths between nodes.
 
 _Last updated: 2026-06-23_
@@ -20,8 +20,8 @@ on different nodes or clusters to reach each other via SRv6-encapsulated traffic
 
 VPC and VPCAttachment CRDs are owned by a separate companion operator
 (`go.miloapis.com/cosmos`). Galactic receives pre-populated identifiers through the
-CNI config and acts on them. `galactic-router` reconciles BGP CRDs from the same
-cosmos API group directly — no gRPC sidecar, no provider CRD lifecycle.
+CNI config and acts on them. `galactic-router` reconciles BGP CRDs
+(`go.datum.net/network`) directly — no gRPC sidecar, no provider CRD lifecycle.
 
 ### SRv6 SID encoding
 
@@ -50,7 +50,7 @@ galactic/
 │   ├── runtime/             # RouterRuntime interface + RuntimeManager
 │   │   ├── gobgp/           # GoBGP RouterRuntime (tenant role)
 │   │   └── frr/             # FRR RouterRuntime stub (fabric role, Phase 2)
-│   ├── model/               # DesiredRouter and family; re-exports cosmos enums
+│   ├── model/               # DesiredRouter and family; re-exports BGP API enums
 │   ├── hash/                # SHA-256 change detection over DesiredRouter
 │   ├── metadata/            # Build-time version info (Version, GitCommit, etc.)
 │   ├── cni/                 # CNI cmdAdd / cmdDel
@@ -190,7 +190,7 @@ On DEL, the result contains only `cniVersion` (empty result is correct since the
 | `internal/runtime`            | galactic-router | `RouterRuntime` interface; `RuntimeManager` (keyed map of live runtimes, double-checked lock create) | Yes (runtime map) |
 | `internal/runtime/gobgp`      | galactic-router | Embeds GoBGP v4; lazy-starts on first Apply; handles peer add/update/delete, EVPN paths, policies; tracks established timestamps | Yes (per-router) |
 | `internal/runtime/frr`        | galactic-router | FRR stub — returns `errNotImplemented` for every method                                             | No         |
-| `internal/model`              | both            | `DesiredRouter`, `DesiredPeer`, `DesiredAdvertisement`, `DesiredPolicy`, `RuntimeStatus`; re-exports cosmos enums | No         |
+| `internal/model`              | both            | `DesiredRouter`, `DesiredPeer`, `DesiredAdvertisement`, `DesiredPolicy`, `RuntimeStatus`; re-exports BGP API enums | No         |
 | `internal/hash`               | galactic-router | SHA-256 fingerprint of `DesiredRouter` for no-op suppression                                        | No         |
 | `internal/metadata`           | both            | Build-time vars (`Version`, `GitCommit`, `GitTreeState`, `BuildDate`) stamped via `-ldflags`         | No         |
 | `internal/cni`                | galactic-cni    | `cmdAdd` / `cmdDel`; CNI PluginConf parsing; BGPAdvertisement lifecycle; delegates kernel work to plumbing | No         |
@@ -208,7 +208,7 @@ On DEL, the result contains only `cniVersion` (empty result is correct since the
 | Dependency                              | Version  | Purpose                                                  |
 |-----------------------------------------|----------|----------------------------------------------------------|
 | `github.com/osrg/gobgp/v4`             | v4.6.0   | Embedded BGP server (tenant role)                        |
-| `go.miloapis.com/cosmos`               | pinned   | Cosmos BGP CRD API types (BGPRouter, BGPPeer, etc.)      |
+| `go.datum.net/network`                  | pinned   | BGP CRD API types (BGPRouter, BGPPeer, etc.)             |
 | `sigs.k8s.io/controller-runtime`       | v0.24.1  | Reconciler framework, manager, field indexes             |
 | `github.com/containernetworking/cni`   | v1.3.0   | CNI plugin spec, skel, invoke                            |
 | `github.com/vishvananda/netlink`        | pinned   | Linux netlink: VRF, veth, SRv6 routes                   |
@@ -225,7 +225,7 @@ On DEL, the result contains only `cniVersion` (empty result is correct since the
 - **USID per endpoint.** Each (VPC, VPCAttachment) pair is assigned a unique /128 USID by the companion operator. The CNI installs an END.DT46 decap route for that /128 and advertises it as the EVPN GWIPAddress. VPC identity is not encoded in the SID.
 - **Base62 interface names.** Kernel interface names use the format `G{9-char-vpc-base62}{3-char-att-base62}{suffix}` (suffix: `V` = VRF, `H` = host veth, `G` = guest veth), fitting in the 14-character kernel limit. The hex form is used for BGP and SRv6; base62 for kernel interfaces.
 - **GoBGP embedded, lazy-started.** GoBGP runs in-process and starts only when the first `BGPRouter` is reconciled (`listenPort=-1`, outbound-only). ASN or RouterID changes trigger a full `Reconfigure` (fresh `BgpServer` — `StopBgp` is not called because it permanently terminates the v4 Serve loop).
-- **CRD-driven config, no sidecar gRPC.** `galactic-router` watches cosmos BGP CRDs directly via controller-runtime. The CNI writes a `BGPAdvertisement` CRD; the router reconciler picks it up. No in-node gRPC calls.
+- **CRD-driven config, no sidecar gRPC.** `galactic-router` watches BGP CRDs directly via controller-runtime. The CNI writes a `BGPAdvertisement` CRD; the router reconciler picks it up. No in-node gRPC calls.
 - **Hash-based no-op suppression.** SHA-256 over the sorted `DesiredRouter` prevents redundant GoBGP Apply calls on every CRD event.
 - **RuntimeFactory pattern.** `GALACTIC_ROUTER_ROUTER_MODE=tenant` selects GoBGP; `GALACTIC_ROUTER_ROUTER_MODE=fabric` selects FRR (Phase 2 stub). The binary is selected at startup; no controller changes are needed for Phase 2.
 - **gRPC health on :5000.** Liveness and readiness probes use the gRPC health protocol (`google.golang.org/grpc/health`) on port 5000. No HTTP health endpoint.
@@ -264,7 +264,7 @@ Triggered by `v*` tags. Builds and pushes `ghcr.io/datum-cloud/galactic:{version
 ## Known Constraints
 
 - **GoBGP RIB is ephemeral.** All BGP state is in-process memory. On restart, sessions and paths must be re-established from CRD state; controller-runtime's reconcile loop handles this automatically.
-- **EVPN Type 5 deferred.** `BGPAdvertisement` does not carry a Route Distinguisher field in the current cosmos API. `galactic-router` returns `ErrMissingRouteDistinguisher` for l2vpn/evpn advertisements and sets `Accepted=False` on the CRD.
+- **EVPN Type 5 deferred.** `BGPAdvertisement` does not carry a Route Distinguisher field in the current BGP API. `galactic-router` returns `ErrMissingRouteDistinguisher` for l2vpn/evpn advertisements and sets `Accepted=False` on the CRD.
 - **No kernel-path unit tests.** `internal/cni`, `internal/plumbing/srv6`, and `internal/plumbing/vrf` require `CAP_NET_ADMIN` and a real kernel. `internal/plumbing/intf` is fully unit-testable (pure functions only). Coverage comes from the e2e suite (`task test:e2e`).
 
 ---
