+++ deploy/containerlab/resources/bgp/tenant/dfw/bgpadvertisement-vpc10.yaml
@@ -1,4 +1,4 @@
-apiVersion: bgp.miloapis.com/v1alpha1
+apiVersion: network.datumapis.com/v1alpha1
 kind: BGPAdvertisement
 metadata:
   name: dfw-worker-vpc10-sid
