+++ deploy/containerlab/resources/bgp/tenant/dfw/bgprouter.yaml
@@ -1,4 +1,4 @@
-apiVersion: bgp.miloapis.com/v1alpha1
+apiVersion: network.datumapis.com/v1alpha1
 kind: BGPRouter
 metadata:
   name: dfw-worker-tenant
