+++ deploy/containerlab/resources/bgp/tenant/iad/bgppeer.yaml
@@ -1,4 +1,4 @@
-apiVersion: bgp.miloapis.com/v1alpha1
+apiVersion: network.datumapis.com/v1alpha1
 kind: BGPPeer
 metadata:
   name: iad-worker-tenant-to-rr
