+++ deploy/containerlab/resources/bgp/tenant/sjc/bgpadvertisement-vpc10.yaml
@@ -1,4 +1,4 @@
-apiVersion: bgp.miloapis.com/v1alpha1
+apiVersion: network.datumapis.com/v1alpha1
 kind: BGPAdvertisement
 metadata:
   name: sjc-worker-vpc10-sid
