+++ internal/cni/ipam_ops.go
@@ -10,11 +10,11 @@ import (
 	"net"
 
 	"github.com/containernetworking/cni/pkg/skel"
-	bgpv1alpha1 "go.miloapis.com/cosmos/api/bgp/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"go.datum.net/galactic/internal/cni/ipam"
+	bgpv1alpha1 "go.datum.net/network/api/v1alpha1"
 )
 
 // configureIPAM allocates a subnet and configures the guest interface inside the
