+++ internal/controller/bgppeer_controller.go
@@ -7,11 +7,12 @@ package controller
 import (
 	"context"
 
-	bgpv1alpha1 "go.miloapis.com/cosmos/api/bgp/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	bgpv1alpha1 "go.datum.net/network/api/v1alpha1"
 )
 
 // BGPPeerReconciler watches BGPPeer resources and enqueues the owning BGPRouter(s).
