+++ internal/controller/indexer.go
@@ -8,9 +8,10 @@ import (
 	"context"
 	"fmt"
 
-	bgpv1alpha1 "go.miloapis.com/cosmos/api/bgp/v1alpha1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	bgpv1alpha1 "go.datum.net/network/api/v1alpha1"
 )
 
 // Field index names used across controllers.
